Moved the lsJpi type aliases to a header guarded include, added guards to ChannelQuantumNumbers
