+++ src/resonanceReconstruction/rmatrix/QuantumNumbers.hpp
@@ -1,0 +1,22 @@
+#ifndef NJOY_R2_RMATRIX_QUANTUMNUMBERS
+#define NJOY_R2_RMATRIX_QUANTUMNUMBERS
+
+// system includes
+
+// other includes
+
+namespace njoy {
+namespace resonanceReconstruction {
+namespace rmatrix {
+
+  // quantum numbers
+  using OrbitalAngularMomentum = unsigned int;
+  using Spin = double;
+  using TotalAngularMomentum = double;
+  using Parity = short;
+
+} // rmatrix namespace
+} // resonanceReconstruction namespace
+} // njoy namespace
+
+#endif
